+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2021-11-04
+
+### Added
+
+- Create a ticket with a certain priority.
+
+### Removed
+
+- Superfluous comments.
+
 ## [0.4.0] - 2021-11-04
 
 ### Added
@@ -36,7 +46,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Issue JQL queries.
 - Describe in README how to use the jops library.
 
-[Unreleased]: https://github.com/030/jops/compare/0.4.0...HEAD
+[Unreleased]: https://github.com/030/jops/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/030/jops/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/030/jops/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/030/jops/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/030/jops/compare/0.1.0...0.2.0